requirements.txt
flask>=3.0.0
huggingface-hub>=0.20.0
sqlparse>=0.4.4
requests>=2.31.0
python-dotenv>=1.0.0
